Move TimestampIndexMseTest & TimestampIndexSseTest as they are not part of the suite.
